feat(dashboard): add writing page and update app route
- added `writing.tsx` page under dashboard features
- updated `app/dashboard/writing/page.tsx` to use new structure
